+++ src/System.Windows.Forms.Design/src/System/Windows/Forms/Design/ControlDesigner.ChildSubClass.cs
@@ -46,7 +46,7 @@ namespace System.Windows.Forms.Design
 
                 if (m.MsgInternal == User32.WM.PARENTNOTIFY && (User32.WM)PARAM.LOWORD(m.WParamInternal) == User32.WM.CREATE)
                 {
-                    _designer.HookChildHandles((HWND)m.LParamInternal); // they will get removed from the collection just above
+                    _designer.HookChildHandles((HWND)(nint)m.LParamInternal); // they will get removed from the collection just above
                 }
 
                 // We want these messages to go through the designer's WndProc method, and we want people to be able
